Fix build on 64-bit and on GCC 14
Three long-standing defects, all surfaced by moOde 10 on RPiOS Trixie 64-bit
(GCC 14, kernel 6.12) and reported in issue #4:

- peppyalsa.c: the config locals read by snd_config_get_integer() were int,
  but that function writes a long. On LP64 it stores 8 bytes into a 4-byte
  slot, corrupting the adjacent locals. Declaring them long fixes the read;
  decay_ms was already long for this reason.

- meter.c / spectrum.c: reader_disconnect_handler() was declared (void) but
  installed as a SIGPIPE handler, whose type is void(*)(int). GCC 14 makes the
  resulting incompatible-pointer-types assignment an error rather than a
  warning. Give the handler the (int signum) it is called with.

- meter.c: the meter init() was defined with four parameters while the device
  init pointer, its call site, and spectrum's init all use eight. Same GCC 14
  incompatible-pointer-types error on _meter.init = &init. Widen the signature
  to match; the extra parameters are unused here and cast to void.

These are the fixes carried in moOde's peppyalsa packaging (patch by kent_reed);
this backports them to the original repository.

+++ src/peppyalsa.c
@@ -251,16 +251,16 @@ int _snd_pcm_scope_peppyalsa_open(
     int err; 
     
     const char *meter_fifo = "";
-    int meter_max = -1;
-    int meter_show = -1;
-    
+    long meter_max = -1;
+    long meter_show = -1;
+
     const char *spectrum_fifo = "";
-    int spectrum_max = -1;
-    int spectrum_size = -1;
-    int log_f = -1;
-    int log_y = -1;
-    int s_factor = -1;
-    int window = -1;
+    long spectrum_max = -1;
+    long spectrum_size = -1;
+    long log_f = -1;
+    long log_y = -1;
+    long s_factor = -1;
+    long window = -1;
 
     num_meters = MAX_METERS;
     num_scopes = MAX_METERS;
